Marcas Controller y Service agregados

+++ tiendaAPI/src/main/java/org/valerio/tiendaapi/repository/MarcasRepository.java
@@ -3,5 +3,5 @@ package org.valerio.tiendaapi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.valerio.tiendaapi.model.Marcas;
 
-public interface MarcasRepository extends JpaRepository<Marcas, Long> {
+public interface MarcasRepository extends JpaRepository<Marcas, Integer> {
 }
